requirements.txt
# Kortex Agent Dependencies
# Versions pinned for reproducible builds

# HTTP & Async
aiohttp==3.13.2

# AI Providers
anthropic==0.75.0
google-generativeai==0.8.5
openai==2.9.0

# Web Framework
Flask==3.1.2
flask-cors==6.0.1

# Document Processing
PyPDF2==3.0.1

# Environment
python-dotenv==1.2.1

# Testing
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
